% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordillera.R
\name{cordillera}
\alias{cordillera}
\title{Calculates the Optics Cordillera}
\usage{
cordillera(X, q = 1, minpts = 2, epsilon, distmeth = "euclidean",
  dmax = NULL, rang, digits = 10, scale = TRUE, ...)
}
\arguments{
\item{X}{numeric matrix or data frame representing coordinates of points, or a symmetric matrix of distance of points or an object of class \link{dist}. Passed to \code{\link{optics}}, see also there.}

\item{q}{the norm of the cordillera. Defaults to 1.}

\item{minpts}{the minpts argument to elki. Defaults to 2.}

\item{epsilon}{The epsilon parameter for OPTICS. Defaults to 2 times the range of x.}

\item{distmeth}{The distance to be computed if X is not a symmetric matrix or a dist object (otherwise ignored). Defaults to Euclidean distance.}

\item{dmax}{The winsorization value for the highest allowed reachability. If used for comparisons this should be supplied. If no value is supplied, it is NULL (default), then dmax is taken from the data as minimum of epsilon or the largest reachability.}

\item{rang}{(DEPRECATED) A range of values for making up dmax. If supplied it overrules the dmax parameter and rang[2]-rang[1] is returned as dmax in the object. If no value is supplied rang is taken to be (0, dmax) taken from the data. Only use this when you know what you're doing, which would mean you're me (and even then we should be cautious).}

\item{digits}{round the raw cordilerra and the norm factor to these digits. Defaults to 10.}

\item{scale}{Should the confs be scaled to mean 0 and sd 1? Defaults to TRUE}

\item{...}{Additional arguments to be passed to optics}
}
\value{
A list with the elements
     \itemize{
       \item $raw... The raw cordillera
       \item $norm... The normalization constant
       \item $normfac... The normalization factor (the number of times that dmax is taken)
       \item $dmaxe... The effective maximum distance used for maximum structure (either dmax or epsilon or rang[2]-rang[1]). 
       \item $normed... The normed cordillera (raw/norm)
       \item $optics... The optics object
}
}
\description{
Calculates the OPTICS cordillera as in Rusch et al., 2015. Based on optics in dbscan package.
}
\section{Warning}{
 It may happen that the (normed) cordillera cannot be calculated properly (e.g. division by zero, infinite raw cordillera, q value to high etc.). A warning will be printed and the normed cordillera is either 0, 1 (if infinity is involved) or NA. In that case one needs to check one or more of the following: reachability values returned from optics, minpts, eps, the raw cordillera, dmax and the normalization factor normfac.
}
\examples{
data(iris)
res<-princomp(iris[,1:4])
#2 dim goodness-of-clusteredness with clusters of at least 2 points
#With a matrix of points
cres2<-cordillera(res$scores[,1:2])
cres2
summary(cres2)
plot(cres2)

#with a dist object 
dl0 <- dist(res$scores[,1:2],"maximum") #maximum distance
cres0<-cordillera(dl0)
cres0
summary(cres0)
plot(cres0)

#with any symmetric distance/dissimilarity matrix 
dl1 <- cluster::daisy(res$scores[,1:2],"manhattan") 
cres1<-cordillera(dl1)
cres1
summary(cres1)
plot(cres1)

#4 dim goodness-of-clusteredness with clusters of at least 3 points for PCA
cres4<-cordillera(res$scores[,1:4],minpts=3,epsilon=13) 
#4 dim goodness-of-clusteredness with clusters of at least 3 points for original data
cres<-cordillera(iris[,1:4],minpts=3,epsilon=13,dmax=cres4$dmaxe)
#There is a bit more clusteredness for the PCA result
summary(cres4)
summary(cres)
plot(cres4)
plot(cres)
}
\keyword{clustering}
\keyword{multivariate}

