% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{plot.opticse}
\alias{plot.opticse}
\title{Plot method for OPTICS results}
\usage{
\method{plot}{opticse}(x, withlabels = FALSE, col = "grey55",
  colna = "grey80", border = graphics::par("bg"), names.arg, ...)
}
\arguments{
\item{x}{an object of class optics}

\item{withlabels}{flags whether point labels should be drawn. Defaults to TRUE.}

\item{col}{the color of the bars of finite/defined reachabilities. Defaults to grey.}

\item{colna}{the color of the bars for the points with infinite undefined reachabilities. Defaults to a lighter grey.}

\item{border}{the color of the bar borders. Defaults to par("bg")}

\item{names.arg}{... The arguments to be passed as names}

\item{...}{additional arguments passed to barplot}
}
\description{
Displays the reachability plot. Points with undefined/infinite minimum reachabilities are colored lighter by default.
}

