% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{summary.opticse}
\alias{summary.opticse}
\title{Summary method for OPTICS results}
\usage{
\method{summary}{opticse}(object, ...)
}
\arguments{
\item{object}{an object of class optics}

\item{...}{additional arguments passed to summary.numeric}
}
\value{
an object of class summary.optics wit the reachabilities, the summary and minpts and epsilon parameters
}
\description{
Displays summaries of the reachability plot. Currently its the five points summary of the reachabilities and a stem and leaf display. The latter should not be confused with the reachability plot. If you need the latter, use plot().
}
